feat(): shuffle-training-set-tuning-validation-split-ratio
(1). Original dataset was organized in longitude order, then when splitting the train and validation sets, the split sets are not generalized for the whole dataset, it is just separated following the longitude pattern. We should shuffle the dataset before splitting to make sure the trained model generalized the pattern. (2). If the validation_split < 0.15, we do not have enough validation examples to feedback for the model which means that the model will be overfitting into the large training dataset. Larger validation_split is good, however, if the ratio is bigger the fewer training examples we will have. The idea validation_split ratio is around 0.2